+++ .config/dotnet-tools.json
@@ -1,0 +1,12 @@
+{
+  "version": 1,
+  "isRoot": true,
+  "tools": {
+    "nexusuploader": {
+      "version": "0.0.6",
+      "commands": [
+        "unex"
+      ]
+    }
+  }
+}
