+++ unex.yml
@@ -1,0 +1,5 @@
+Game: valheim
+ModId: 0
+FileName: ExampleMod
+FileDescription: |-
+  Description for your mod
